Shooter subsystem test

+++ src/Robot.h
@@ -82,8 +82,8 @@ private:
 	AutonomousRoutine m_autoRoutine;
 	uint32_t 					m_autoTimer;
 	int						m_speedSetpt;
-	int						m_flailSetpt;
-	int						m_conveyorSetpt;
+	double						m_flailSetpt;
+	double						m_conveyorSetpt;
 	/**
 	 * Logging
 	 */
